requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
notion-client==2.2.1
anthropic==0.39.0
requests==2.31.0
feedparser==6.0.10
schedule==1.2.2
python-dotenv==1.0.1
arxiv==2.1.0
PyPDF2==3.0.1
